\name{elecyears}
\docType{data}
\alias{elecyears}
\title{ICPSR Data Set 6311: U.S. House Elections, 1896-1992 -
  Election Years }
\description{
This data set augments data on the elections to the U.S. House of
Representatives, every two years between 1896 and 1992.}
\usage{elecyears}
\format{A vector of 49 elements, containing biennial election years from
 1896 to 1992.}
\references{ICPSR.}
\keyword{datasets}
