\name{same.dists}
\docType{data}
\alias{same.dists}
\title{ICPSR Data Set 6311: U.S. House Elections, 1896-1992 -
  Redistricting Indication}
\description{
This data set augments data on the elections to the U.S. House of
Representatives, every two years between 1896 and 1992.}
\usage{same.dists}
\format{A vector of 49 elements, containing 1 if the election uses the
  same district map as its predecessor, 0 otherwise.}
\references{ICPSR.}
\keyword{datasets}
